% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hoge}
\alias{hoge}
\alias{hoge1}
\alias{hoge2}
\title{Hoge() series}
\usage{
hoge(x)

hoge1(x)

hoge2(x)
}
\arguments{
\item{x}{something argument}
}
\description{
There are lots of hoge() s.
}
\section{Functions}{
\itemize{
\item \code{hoge1}: I'm hoge1

\item \code{hoge2}: I'm hoge2
}}

